add_location

+++ src/main/java/ru/vlpetko/weatherbot/model/CurrentWeather.java
@@ -18,11 +18,11 @@ public class CurrentWeather {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long currentId;
 
-    private float temperature;
+    private double temperature;
 
-    private float windSpeed;
+    private double windSpeed;
 
-    private float windDirection;
+    private double windDirection;
 
     private LocalDateTime requestDate;
